fix: deconstruct parser into routines
